currency_api_project copie/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.0.14
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mssql-django==1.5
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
